backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
narwhals==1.27.1
numpy==2.0.2
openai==1.63.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.2
PyMuPDF==1.25.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
streamlit==1.42.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
twilio==9.4.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.2
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
